refactor: split parser constants
